Initial integration between parser and app.
The Parse(ctx) function is now implemented by calling the generated parser. The body of the function is in conditions.l.

Moved the definition of the ParserContext structure into parse.h, since it now needs to be shared between two files.

Made the SetBuiltInVariables function public - it didn't really belong inside Parse().