Rework the grammar to make the rules type-able.
Also decided to remove the <, <=, > and >= operators for now. We'll see how far we get with just = and !=.

Changed the meaning of the 'not' keyword: it's now a synonym for '!=' instead of negating an arbitrary boolean expression. This is part of the Cunning Plan to forcibly keep things simple in .deps files.