Simplify the grammar and start adding actions to it.
Removed ints from the grammar (for now). Focus is on getting it working for strings first, then will go and add support for ints back in.

Added %type declarations for each of the rules.

Added semantic actions to all rules. Some actions are just placeholders at the moment, but many of them are working.